+++ README.md
@@ -292,6 +292,35 @@ heroku 배포 링크 : <https://java-spring-cafe.herokuapp.com>
 </details>
 
 
+<br/>
+<details>
+<summary>🏄‍♂️ 7단계 요구사항</summary>
+<div markdown="1">
+<br/>
+
+## 스프링 카페 6단계 - 게시글 권한 부여
+
+### 구현 요구사항
+
+- [ ] RestController 와 AJAX로 구현한다.
+- [ ] 데이터베이스는 MySQL 8.0을 이용한다.
+
+### 배포 요구사항
+
+- [ ] AWS EC2로 배포한다.
+- [ ] 배포 IP를 README에 기술한다.
+- [ ] 데이터베이스와 같은 인스턴스를 사용한다.
+
+### 추가 요구 사항
+
+- [X] 새로운 답변을 추가하는 기능을 AJAX로 구현한다.
+- [ ] 답변 삭제하는 기능을 AJAX로 구현한다.
+
+<br/>
+</div>
+</details>
+
+
