+++ build.gradle
@@ -21,6 +21,9 @@ dependencies {
 	// h2 database
 	implementation group: 'org.springframework.boot', name: 'spring-boot-starter-jdbc'
 	implementation group: 'com.h2database', name: 'h2', version: '2.0.206'
+
+	// dev-tools
+	implementation 'org.springframework.boot:spring-boot-devtools'
 }
 
 // https://docs.gradle.org/current/userguide/jacoco_plugin.html#header
