+++ src/main/java/com/kakao/cafe/controller/ArticleController.java
@@ -3,7 +3,6 @@ package com.kakao.cafe.controller;
 import com.kakao.cafe.dto.ArticleResponse;
 import com.kakao.cafe.dto.ArticleSaveRequest;
 import com.kakao.cafe.service.ArticleService;
-import com.kakao.cafe.service.ReplyService;
 import com.kakao.cafe.session.SessionUser;
 import java.util.List;
 import javax.servlet.http.HttpSession;
@@ -19,11 +18,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 public class ArticleController {
 
     private final ArticleService articleService;
-    private final ReplyService replyService;
 
-    public ArticleController(ArticleService articleService, ReplyService replyService) {
+    public ArticleController(ArticleService articleService) {
         this.articleService = articleService;
-        this.replyService = replyService;
     }
 
     // question
@@ -76,22 +73,6 @@ public class ArticleController {
         HttpSession session) {
         SessionUser user = SessionUser.from(session);
         articleService.deleteArticle(user, articleId);
-        return "redirect:/";
-    }
-
-    // reply
-    @PostMapping("articles/{id}/answers")
-    public String createAnswer(@PathVariable(value = "id") Integer articleId,
-        String comment, HttpSession session) {
-        SessionUser user = SessionUser.from(session);
-        replyService.comment(user, articleId, comment);
-        return "redirect:/";
-    }
-
-    @DeleteMapping("articles/{questionId}/answers/{id}")
-    public String deleteAnswer(@PathVariable(value = "id") Integer replyId, HttpSession session) {
-        SessionUser user = SessionUser.from(session);
-        replyService.deleteReply(user, replyId);
         return "redirect:/";
     }
