+++ src/main/resources/templates/components/answerTemplate.html
@@ -1,0 +1,30 @@
+<script type="text/template" id="answerTemplate">
+  <article class="article">
+    <div class="article-header">
+      <div class="article-header-thumb">
+        <img src="https://graph.facebook.com/v2.3/1324855987/picture" class="article-author-thumb"
+             alt="">
+      </div>
+      <div class="article-header-text">
+        <a href="#" class="article-author-name">{0}</a>
+        <div class="article-header-time">{1}</div>
+      </div>
+    </div>
+    <div class="article-doc comment-doc">
+      {2}
+    </div>
+    <div class="article-util">
+      <ul class="article-util-list">
+        <li>
+          <a class="link-modify-article" href="#">수정</a>
+        </li>
+        <li>
+          <form class="delete-answer-form" action="/questions/{3}/answers/{4}" method="POST">
+            <input type="hidden" name="_method" value="DELETE">
+            <button type="submit" class="delete-answer-button">삭제</button>
+          </form>
+        </li>
+      </ul>
+    </div>
+  </article>
+</script>
